v0.6-Agents-Code-Structure modified:   README.md new file:   dev_notes/GUI-Welcome - PyTermGUI__docs.pdf renamed:    dev_notes/startGUI.py -> dev_notes/GUI-startGUI.py renamed:    dev_notes/tui-rich-example.py -> dev_notes/GUI-tui-rich-example.py renamed:    dev_notes/File-Overview.md -> dev_notes/Prompt-File-Overview.md new file:   dev_notes/Python-Factory.md renamed:    dev_notes/env-variables-and-security.md -> dev_notes/Python-env-variables-and-security.md renamed:    dev_notes/new-project-file-structure.md -> dev_notes/Python-new-project-file-structure.md renamed:    dev_notes/execution_flow.png -> dev_notes/Sketch-execution-flow.png renamed:    dev_notes/header.py -> dev_notes/Template-Python-header.py renamed:    dev_notes/header.md -> dev_notes/Template-md-header.md new file:   dev_notes/llm-client/.vscode/PythonImportHelper-v2-Completion.json deleted:    dev_notes/v0.1-FINAL.rar deleted:    dev_notes/v0.3_FINAL.rar new file:   dev_notes/v0.6-Agents-Code-Structure new file:   dev_notes/v0.6-Current-Agent-Role-Refine.md new file:   dev_notes/v0.6-New-Agents-Idea.md deleted:    log/log_crew-2024-12-27_00-15-33.md deleted:    log/log_crew-2024-12-27_17-53-15.md new file:   log/log_crew-2025-01-05_13-25-50.md new file:   log/log_prompt_master-2025-01-05_13-25-50.md deleted:    log/log_researcher-2024-12-29_08-26-04.md new file:   log/log_researcher-2025-01-05_13-25-50.md new file:   log/log_writer-2025-01-05_13-25-50.md modified:   main.py deleted:    podcast/Podcast_ An-Orwellian-AI-News-Analyzer.mp3 renamed:    screenshot/PySimpleGUI.jpg -> screenshots/PySimpleGUI.jpg renamed:    screenshot/Sketch-TUI.jpg -> screenshots/Sketch-TUI.jpg renamed:    screenshot/TUI-First-run.jpg -> screenshots/TUI-First-run.jpg renamed:    screenshot/TUI-Good.jpg -> screenshots/TUI-Good.jpg renamed:    screenshot/env_manager_menu_Flowchart-sketch.jpg -> screenshots/env_manager_menu_Flowchart-sketch.jpg renamed:    screenshot/sketchy-ui.png -> screenshots/sketchy-ui.png renamed:    showcase/README.md -> showcases/README.md renamed:    showcase/Showcase-1/Prompt1-ChatGPT.jpg -> showcases/Showcase-1/Prompt1-ChatGPT.jpg renamed:    showcase/Showcase-1/Prompt1.Dall-E-2.jpg -> showcases/Showcase-1/Prompt1.Dall-E-2.jpg renamed:    showcase/Showcase-1/Prompt1.Dall-E-3.jpg -> showcases/Showcase-1/Prompt1.Dall-E-3.jpg renamed:    showcase/Showcase-1/Prompt1.Dall-E-4.jpg -> showcases/Showcase-1/Prompt1.Dall-E-4.jpg renamed:    showcase/Showcase-1/Prompt1.Dall-E.jpg -> showcases/Showcase-1/Prompt1.Dall-E.jpg renamed:    showcase/Showcase-1/Prompt2-ChatGPT.jpg -> showcases/Showcase-1/Prompt2-ChatGPT.jpg renamed:    showcase/Showcase-1/Prompt2-Dall-E-2.jpg -> showcases/Showcase-1/Prompt2-Dall-E-2.jpg renamed:    showcase/Showcase-1/Prompt2-Dall-E-3.jpg -> showcases/Showcase-1/Prompt2-Dall-E-3.jpg renamed:    showcase/Showcase-1/Prompt2-Dall-E-4.jpg -> showcases/Showcase-1/Prompt2-Dall-E-4.jpg renamed:    showcase/Showcase-1/Prompt2-Dall-E.jpg -> showcases/Showcase-1/Prompt2-Dall-E.jpg renamed:    showcase/Showcase-1/terminal_log.md -> showcases/Showcase-1/terminal_log.md renamed:    showcase/Showcase-2/Prompt1-ChatGPT.jpg -> showcases/Showcase-2/Prompt1-ChatGPT.jpg renamed:    showcase/Showcase-2/Prompt1-Dalle-E-2.jpg -> showcases/Showcase-2/Prompt1-Dalle-E-2.jpg renamed:    showcase/Showcase-2/Prompt1-Dalle-E-3.jpg -> showcases/Showcase-2/Prompt1-Dalle-E-3.jpg renamed:    showcase/Showcase-2/Prompt1-Dalle-E-4.jpg -> showcases/Showcase-2/Prompt1-Dalle-E-4.jpg renamed:    showcase/Showcase-2/Prompt1-Dalle-E.jpg -> showcases/Showcase-2/Prompt1-Dalle-E.jpg renamed:    showcase/Showcase-2/Prompt2-ChatGPT.jpg -> showcases/Showcase-2/Prompt2-ChatGPT.jpg renamed:    showcase/Showcase-2/Prompt2-Dalle-2.jpg -> showcases/Showcase-2/Prompt2-Dalle-2.jpg renamed:    showcase/Showcase-2/Prompt2-Dalle-3.jpg -> showcases/Showcase-2/Prompt2-Dalle-3.jpg renamed:    showcase/Showcase-2/Prompt2-Dalle-4.jpg -> showcases/Showcase-2/Prompt2-Dalle-4.jpg renamed:    showcase/Showcase-2/Prompt2-Dalle.jpg -> showcases/Showcase-2/Prompt2-Dalle.jpg renamed:    showcase/Showcase-2/terminal-output--26-12-2024.md -> showcases/Showcase-2/terminal_log2.md new file:   src/agents.py modified:   src/config.py modified:   src/llm.py
