+++ dev_notes/GUI-tui-rich-example.py
