+++ dev_notes/v0.6-Agents-Code-Structure
@@ -1,0 +1,7 @@
+# George_Was_Right v0.6 - New Agent Code Structure
+
+## Idea
+
+```mermaid
+
+```
