--- log/log_crew-2024-12-27_17-53-15.md
@@ -1,7 +1,0 @@
-2024-12-27 17:53:27: task_name="Researcher Task", task="Search for recent real world news that demonstrate how Orwell's book '1984' is still relevant today1. Begin by identifying key themes from Orwell's '1984' such as surveillance, censorship, propaganda, manipulation of truth, and authoritarian control.
-2. Use the provided SerperDevTool to search the internet for recent news events that align with these themes. For example, use search queries like 'modern surveillance news 2023,' 'censorship in media 2023,' 'propaganda in politics 2023,' 'manipulation of truth in news 2023,' or 'authoritarian regimes 2023.'
-3. Analyze the search results to identify news articles or events that strongly resonate with the themes of '1984.'
-4. Select at least 3-5 relevant news articles that demonstrate the relevance of '1984' in today's world.
-5. Compile a report summarizing the selected news articles, highlighting the connection between each event and the themes from '1984.' Ensure the report is clear, concise, and provides direct examples of the parallels.
-6. Save the findings in a structured format to share with the Writer for the next task.", agent="Researcher", status="started"
-
