--- log/log_researcher-2024-12-29_08-26-04.md
@@ -1,1 +1,0 @@
-This report highlights recent real-world news events that demonstrate how Orwell's book '1984' is still relevant today, with themes of surveillance, censorship, propaganda, authoritarianism, and thought control continuing to impact modern society.
