+++ main.py
@@ -14,120 +14,18 @@ from src.config import (
 )
 from src.tools import search_tool
 from src.llm import LLMFactory
+from src.agents import (
+    ResearcherAgent,
+    WriterAgent,
+    PromptMasterAgent,
+    create_tasks,
+)
 import subprocess
 
 # Load environment variables
 load_dotenv()
 
-class ResearcherAgent(Agent):
-    def __init__(self, search_tool=None, **kwargs):
-        if not search_tool:
-            raise ValueError("search_tool is required for ResearcherAgent")
-
-        super().__init__(
-            role="Researcher",
-            goal=(
-                "Finding recent real world news that demonstrate how Orwell's book "
-                "'1984' is still relevant in our days."
-            ),
-            backstory=(
-                "You are an expert researcher with a keen eye for current events "
-                "and literary analysis."
-            ),
-            allow_delegation=True,
-            verbose=True,
-            llm=LLMFactory.init_llm(),
-            tools=[search_tool],
-            **kwargs,
-        )
-
-
-class WriterAgent(Agent):
-    def __init__(self, **kwargs):
-        super().__init__(
-            role="Writer",
-            goal=(
-                "Gather examples of real world news events that could be from "
-                "Orwell's '1984' book"
-            ),
-            backstory=(
-                "You are a skilled writer with a deep understanding of literature "
-                "and current affairs."
-            ),
-            allow_delegation=True,
-            verbose=True,
-            llm=LLMFactory.init_llm(),
-            **kwargs,
-        )
-
-
-
-class PromptMasterAgent(Agent):
-    def __init__(self, **kwargs):
-        super().__init__(
-            role="Prompt Master",
-            goal=("Create visual concepts based on provided prompts"),
-            backstory=(
-                "You are a talented illustrator with a knack for translating ideas "
-                "into compelling visuals."
-            ),
-            allow_delegation=False,
-            verbose=True,
-            llm=LLMFactory.init_llm(),
-            **kwargs,
-        )
-
-
-
-def create_tasks(researcher, writer, prompt_master):
-    return [
-        Task(
-            name="Researcher Task",
-            description=(
-                "Search for recent real world news that demonstrate how Orwell's book "
-                "'1984' is still relevant today"
-            ),
-            agent=researcher,
-            expected_output=(
-                "A report containing recent events that are linked to the themes "
-                "described in Orwell's book, '1984'"
-            ),
-            output_file=log_researcher,
-        ),
-        Task(
-            name="Writer Task",
-            description=(
-                "Compare the news events with themes from '1984' and select the most "
-                "relevant match to write a small article on the subject."
-            ),
-            agent=writer,
-            expected_output=(
-                "A detailed comparison of a selected news event and a theme from "
-                "'1984'"
-            ),
-            output_file=log_writer,
-        ),
-        Task(
-            name="Prompt Master Task",
-            description=(
-                "Create two sets of prompts: one for an illustration for a recent "
-                "event and another one for a linked theme of the book '1984'."
-            ),
-            agent=prompt_master,
-            expected_output=(
-                "Two sets of prompts that will be used to generate illustrations: one "
-                "for the recent event and one for the '1984' book theme."
-            ),
-            output_file=log_prompt_master,
-        ),
-    ]
-
-
-
 def create_crew():
-    try:
-        researcher = ResearcherAgent(search_tool=search_tool)
-        writer = WriterAgent()
     try:
         researcher = ResearcherAgent(search_tool=search_tool)
         writer = WriterAgent()
@@ -145,12 +43,13 @@ def create_crew():
         print(f"Error creating crew: {str(e)}")
         raise
 
-
-
 def main():
     try:
-        validate_environment()
-
+        # Ensure the log directory exists
+        log_directory = os.path.dirname(log_crew)
+        if not os.path.exists(log_directory):
+            os.makedirs(log_directory)
+        
         print(
             f"""
             -------------------
@@ -160,12 +59,6 @@ def main():
             -------------------
             """
         )
-
-        # Execute start.py and capture its output
-        start_output = subprocess.run(
-            ["python", "start.py"], capture_output=True, text=True
-        ).stdout
-        print(start_output)
 
         crew = create_crew()
         crew.kickoff()
@@ -182,7 +75,6 @@ def main():
     except Exception as e:
         print(f"Error running application: {str(e)}")
         raise
-
 
 if __name__ == "__main__":
     main()
