+++ screenshots/PySimpleGUI.jpg
