+++ showcases/Showcase-1/terminal_log.md
