+++ showcases/Showcase-2/terminal_log2.md
